04-api/requirements.txt
fastapi
gunicorn
hvac
psycopg2
pulsar-client==2.5.2
uvicorn
